Change Covid API by CryptoCurrency API
