+++ src/redux/configureStore.js
@@ -1,10 +1,10 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import countriesReducer from './countries/countries';
+import coinsReducer from './coins/coins';
 
 const reducer = combineReducers({
-  countriesReducer,
+  coinsReducer,
 });
 
 const store = createStore(reducer, applyMiddleware(logger, thunk));
